# static/openapi_skg-if_cessda.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: SKG-IF OpenAPI - compatible with SKG-IF Data Model 1.1.0
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    Below the line is the original SKG-IF OpenAPI documentation that has only been modified to reflect the implementation status of CESSDA's endpoints. Endpoints that are not available have a text saying **"Not yet implemented"** next to the endpoint name. Filters that are not applicable / not implemented are marked with a strikethrough.

    --------------------------------------

    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at: https://skg-if.github.io/api/

    # Note for implementers :
      - To validate a live implementation server see: [.github/WORFLOW.md](https://github.com/skg-if/api/blob/main/.github/WORFLOW.md)
      - Please refer to the [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewers:
...
paths:
...
  '/topics/{local_identifier}':
    get:
      tags:
          - Topic
      summary: Get topic by id
      description: |
          Get single `topic`. See definition in SKG-IF [Topic](https://skg-if.github.io/interoperability-framework/docs/topic.html) ( entity_type:topic ).
      operationId: getTopicById
      parameters:
        - $ref : '#/components/parameters/localIdPathParam'
      responses:
        '200':
          description: Success
          content:
            # not 'application/json-ld' to be compatible with StopLight PRISM tool
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/MetaSingleEntity'
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Topic"
              examples:
                TopicEx01:
                  summary: Computer Science topic
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}
                    ]
                    "meta" :
                      local_identifier: https://example.com/skg-if/api/topics/topc-1-cs
                      entity_type: "single_entity"
                    "@graph" :
                      - local_identifier: topic-1-cs
                        entity_type: "topic"
                        identifiers:
                          - scheme: wikidata
                            value: https://www.wikidata.org/wiki/Q21198
                        labels:
                          en: "Computer Science"
                          it: "Informatica"
        '404':
          $ref: '#/components/responses/NotFoundEntity'
  '/topics':
    get:
      tags:
          - Topic
      summary: Get list of topic
      operationId: getTopics
      description: |
          Get a list of `topic`. See definition in SKG-IF [Topic](https://skg-if.github.io/interoperability-framework/docs/topic.html) ( entity_type:topic ).

          __Attribute filters__ \
          You can filter using these attributes of the `topic` object. Exact match.

          | Filter key | Attribute filter example |
          | -------- | -------- |
          | identifiers.scheme | `identifiers.scheme:wikidata` |
          | identifiers.value | `identifiers.value:https://www.wikidata.org/wiki/Q42` |

          __Convenience filters__ \
          These filters (`cf.*`) aren't attributes of the `topic` object, but they're handy for solving some common use cases.

          | Filter key | Filter value | List returned | Attribute filter example |
          | -------- | -------- | ------- | ------- |
          | cf.search.labels | a string | `topic` with label containing the filter value | `cf.search.labels:Solar` |
          | cf.search.language | a string | `topic` with language key equals to the filter value | `cf.search.language:en` |

           __Query parameter syntax examples__ \

           * `/topics?filter=cf.search.labels:Solar,identifiers.scheme:myvocabulary`
           * `/topics?filter=cf.search.labels:Solar`
           * `/topics?filter=cf.search.labels:Solar,cf.search.language:en`
           * `/topics?filter=cf.search.labels:Solar,cf.search.language:en&page=1&page_size=5`

      parameters:
        - name: filter
          in: query
          description: |
                Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
          schema:
            type: string
            pattern: '^(,?.+:.+)*$'
          examples : # note example with parameter $ref. OK on Redoc, KO on stoplight
            topic_label_ex:
              value: cf.search.labels:Solar
              summary: Get all topics with label containing 'Solar'
            topic_label_lang_ex:
              value: cf.search.labels:Solar,cf.search.language:en
              summary: Get all topics with label containing 'Solar' and english
            topic_label_vocab_ex:
              value: cf.search.labels:Solar,identifiers.scheme:myvocabulary
              summary: Get all topics with label containing 'Solar' in the scheme 'myvocabulary'
        - $ref : '#/components/parameters/pageQueryParam'
        - $ref : '#/components/parameters/pageSizeQueryParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/MetaSearch'
                  "@graph":
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/Topic'
                        - $ref: '#/components/schemas/LocalIdentifierRef'
                required: [ "@context", "meta", "@graph" ]
        '404':
          $ref: '#/components/responses/NotFoundSearch'
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    localIdPathParam:
      name: local_identifier
      in: path
      description: |
        Entity id.
        - Example entity id with as URL : `http://skg-base-domain.com/prod-c66c6-38be-4d5f-85db-d44c9`. To define your SKG base domain ( `@base` ), please refer to the [SKG-IF API doc](https://skg-if.github.io/api/)
        - Example entity id without its SKG base domain : `prod-c66c6-38be-4d5f-85db-d44c9`

        The entity endpoint MUST resolve ids as full URL and SHOULD resolve ids without SKG base domain.
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
    pageQueryParam:
      name: page
      in: query
      description: |
            Search page reference (page number or search iteration token).\
            Note: As a client application you should build your iterating logic using the `meta.next_page.local_identifier` and `meta.prev_page.local_identifier` URLs
      schema:
        type: string
    pageSizeQueryParam:
      name: page_size
      in: query
      description: |
            Search page size. limit search result size.
      schema:
        type: integer
  responses:
    NotFoundEntity:
      description: |
        No entity found for the given id. The `@graph` array is empty.
      content:
        application/json:
          schema:
            required: ['@context','@graph']
            properties:
              "@context":
                type: array
                description: JSON-LD context # not displayed in Stoplight elements
                minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                uniqueItems: true
                items:
                  anyOf:
                    - $ref: '#/components/schemas/JsonLdCtxDataModel'
                    - $ref: '#/components/schemas/JsonLdCtxSearch'
                    - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                examples:
                  - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}]
              meta:
                type: object
                $ref: '#/components/schemas/MetaSingleEntity'
              "@graph":
                type: array
                maxItems: 0
                items: {}
          examples:
            NotFoundEntityEx01:
              summary: No entity found for the given id
              value:
                "@context": [
                  "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                  "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                  {"@base" : "https://w3id.org/skg-if/sandbox/acme/"}
                ]
                "meta":
                  local_identifier: https://example.com/skg-if/api/products/unknown-id
                  entity_type: "single_entity"
                "@graph": []
...
  schemas:
...
    ApiItem:
      type: object
      title: 'Link'
      description:  |
         Result element
      required: [
         "local_identifier"
      ]
      properties:
        local_identifier:
          type: string
          description: |
            Unique local identifier as URL
          examples:
            - http://w3id.org/skg-if/sandbox/acme/product-1
            - http://w3id.org/skg-if/sandbox/acme/person-1
        urls:
          type: [array, "null"]
          description: |
            List of API URLs to access to the entity metadata
          items:
            $ref: '#/components/schemas/Link'
...
    Topic:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          title: 'Topic'
          description:  "`topic` object containing a limited set of fields.  See definition in SKG-IF [Topic](https://skg-if.github.io/interoperability-framework/docs/topic.html) ( entity_type:topic )."
          properties :
            entity_type:
              type: string
              enum:
                - "topic"
              x-faker:
                fake: ["topic"]
            identifiers:
              description: Objects representing external identifiers for the entity. See [Topic](https://skg-if.github.io/interoperability-framework/docs/topic.html#identifiers) documentation for supported schemes
              type: [array, "null"]
              items:
                required: [ "scheme", "value" ]
                properties:
                  scheme:
                    description: The scheme for the external identifier.
                    type: string
                    #TODO regexp no space
                    examples:
                      - wikidata
                      - dbpedia
                    x-faker:
                      helpers.arrayElement: [["wikidata", "dbpedia", "computer_science_ontology"]]
                  value:
                    type: string
                    examples:
                      - https://www.wikidata.org/wiki/Q21198
                    x-faker:
                      fake : ['{{internet.url}}/topic_{{random.alphaNumeric(6)}}']
            labels:
              type: [object, "null"]
              description:  The labels describing a Topic (multiple for multilingualism). The object is a dictionary, the keys represent language codes following ISO 639-1 (`en`, `es`, `it`, `fr`...). The special key `none` is reserved whenever the information about the language is not available or cannot be shared.
              patternProperties:
                "^([a-z]{2}|none)$":
                  type: string
                  x-faker:
                    random.words: [1]
              example: # 'object' type only supports single value 'example' ( not 'examples' ) https://swagger.io/docs/specification/v3_0/adding-examples/
                en: "Computer Science"
                it: "Informatica"
...
    MetaSingleEntity:
      type: object
      required: [ "local_identifier", "entity_type" ]
      properties:
        local_identifier:
          type: string
          description: Entity API URL. local_identifier can be expressed with ot without full domain. Full domain, including your @base, MUST also resolve
          examples:
            - https://example.com/skg-if/api/products/product-1
            - https://example.com/skg-if/api/persons/person-1
            - https://acme.com/skg-if/api/products/https://w3id.org/skg-if/sandbox/acme/prod-1
            - https://acme.com/skg-if/api/persons/https://w3id.org/skg-if/sandbox/acme/person-1
        entity_type:
          type: string
          description: single entity
          enum:
            - single_entity
        api_items:
          type: [array, "null"]
          items:
            $ref: '#/components/schemas/ApiItem'
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/acme" }`'
      properties:
        "@base":
          type: [string, "null"]
          description : |
            Default root URL of the JSON-LD context. \
            To define your `@base`, please refer to the [SKG-IF API doc](https://skg-if.github.io/api/) \
            Your `@base` should end with a `/` \
            common formats : \
            - `https://acme.com/skg/` \
            - `https://w3id.org/acme/` \
            - `https://w3id.org/skg-if/sandbox/acme/`\
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://acme.com/skg/"
            - "https://acme.com/graph/"
            - "https://w3id.org/acme/"
            - "https://w3id.org/skg-if/sandbox/acme/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg/{{random.alpha(3)}}']